Add more tests.
The method which was responsible for filtering collections names has been removed from the Collection class
and integrated into the dataprocess module because it also works very well for movies.

So, Movie and Collection classes now use the same code to filter their name, and pass the tests.
An option allowing to set a maximum number of characters has also been added.