Change drag and drop management.
Previously, poster drag and drop only worked when a movie was selected.
Now, it is no longer necessary to select a movie, drag and drop modifies the image of the movie
currently displayed even if not selected.
This is possible thanks to the introduction of the last_movie_displayed class attribute.